Reboot - Timer.ps1
Use this script to reboot a Windows device with a variable to submit how many minutes (not seconds) before the reboot happens with a custom notification to the end user. Too many users forgetting to submit "seconds" instead of minutes, causing 5-second reboots, for example. This script solves that easily made mistake.